+++ test/font_test_data.dart
@@ -1,0 +1,3 @@
+export 'font_test_data_stub.dart'
+    if (dart.library.io) 'font_test_data_io.dart'
+    if (dart.library.js_interop) 'font_test_data_web.dart';