+++ test/font_test_data_stub.dart
@@ -1,0 +1,5 @@
+import 'dart:typed_data';
+
+String? get systemFontFamily => null;
+
+Future<Uint8List?> loadTinosFontData() async => null;